user_portal/requirements.txt
auth==0.5.3
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
crispy==0.6.3
cryptography==3.2.1
cycler==0.10.0
dj-rest-auth==2.1.10
Django==2.2.24
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
dnspython==1.16.0
eventlet==0.31.1
fabio==0.12.0
falcon==3.0.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.1.1
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
mongoengine==0.22.0
numpy==1.18.5
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.12.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.1
serializers==0.2.4
silx==0.15.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
uWSGI==2.0.19.1
Werkzeug==1.0.1
zipp==1.2.0
